New Character Animations
Fixed idle animation, added a running animation, and re-did the character sprite sheet. The jumping and falling animations are not fully implemented, and thus, appear slightly broken. Sprites for a wall slide animation are available, but the animation itself is not implemented. I deleted the obsolete player and multiplayer_player, as well as the old test level.