Still Water is Here! (merge #126 first pls)
I added 5 deadly still water scenes that can be used as tiles. They are different only in shape, not in function: for example, still_water_right has a tile that has decorative ground sloping into the tile to its right. To test these changes more easily, and to provide a way to more easily test new obstacles in the future, I created a test_player_character that does not wait for a timer before starting and can be reset if the user presses "R". I also added a test_level so that this and future obstacles could be tested easily before they are implemented into the main game. I lowercased the "d" on line 355 in player_character.gd, "anim_player.play("Dead")," because the name of the animation "dead" in the animation tree is lowercased. Finally, I created a new folder under "scenes" called "obstacles," and I put the still_water variants and the desert_spikes into it for organization.