Adding A Hud
A hud for the players has been added in. Players can now see see when they can dash on the fuel level on their current powerup.

Displaying the powerup on the Hud has been a bit tricky with the current way we are updating the powerup.

Changed the fuel system to where the dash powerup fuel and jetpack fuel share the same fuel variable.