Breakable Wall and Fake Floor are here!
I imported the fake floor obstacle that I made for a previous sprint. I also imported and finished the destroyable wall obstacle that I was working on before the super smash bolt update. After hours of relentless debugging, the wall works for multiplayer... mostly. Sometimes, when a player destroys two walls, only one is shown as destroyed for other players. I don't know what triggers this behavior, so I'm not going to worry about it right now. Additionally, I put sprite assets for the desert spikes, transient platform, and transient wall into one folder, and I buried the animated_sprites into the swamp_objects folder, since that is where they seemed to belong. I imported 2 chunks from the previous version of the game that I had made to show off the fake floor, and I made one brand new chunk to show off the destroyable wall. I updated the player_character to check that when it is looking for something to collide into while dashing, it makes sure it is another player before bumping it, because this was causing issues with the destroyable wall. Aside from a couple tweaks to the test_player_character, the rest are cache updates.