Finalized Animations
Incorporated a dash animation and a jumping animation, and with leftover sprites I created transitions between those animations. The powerup dash still just defaults to the running animation currently. I reworked the collision checking for the get_bumped() so that an entity in the player's collision has its script compared to the name of the player's script, which should ensure that get_bumped() gets called on nothing without a matching script. Namely, it should only get called if players are colliding and at least one is dashing.

I fixed the problem of the dash effect being too low, but it is are still to the right of the character.

Since the spritesheet is complete (I think), I made some recolored versions of the robot. There are now 8 colors of robot to play as. (I did make one robot spritesheet inaccessible because its color scheme did not match the background very well, so there are really 7 colors to be.) I have them randomly assigned to the players right now. They still do not meet Dr. Geary's expectation of being correctly colored upon being loaded in, but I have an idea on how to make it appear this way.

The old mothersheets are obsoleted, deleted, and replaced, but the multiplayer menu still has dependencies on the other old character sprite sheets.

A death sprite is in each of the spritesheets, even though there is no death animation yet. I know what I'd like to do with that, but that's for a later commit.