Split Fuel HUD
Upon activating a powerup that uses fuel, the UI displays an extra battery to show how much fuel is left for the powerup, and the baseplate of the UI is extended. Normal dash fuel is still displayed in its regular battery. Once the powerup fuel source is depleted, the extra battery and extended UI base disappear.