Added Icon and Purpled the Water
Note! The icon is not automatically applied upon export. You need a program called rcedit downloaded and configured to work with Godot to apply the icon. The size variants are in the folder because the console was giving warnings when I only gave it one size. It shouldn't matter which one you choose upon export.

I changed the still water to be purple so that it would look like more threatening, like you might contract arthritis if you stepped in it.