Character Customization Added
Now, one can choose which color to be in the main menu by clicking "Change Color."

I tried to implement player colors in the waiting lobbies, but that requires server changes, and I have spent too much time on this unapproved feature. So, I hid the robot images from the lobby altogether.

I made the little animated robot in the main menu into his own scene so that he could be re-used elsewhere (like the lobby), and so that his color would be easy to change.

I removed randomness to sprite assignment completely, as well as the check for players named "Zorro."